architecture: canonicalize domain state models
