+++ kdn_server/contracts/knowledge.py
@@ -3,7 +3,7 @@ from typing import Literal
 
 from pydantic import Field, model_validator
 
-from kdn_server.domain import CacheArtifact
+from cacheroute.cache import CacheArtifact
 from cacheroute.contracts.v1.common import VersionedMessage
 from cacheroute.contracts.v1.errors import ContractError, OutcomeCode
