+++ kdn_server/gateway/mock.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 from kdn_server.contracts.cache_service import *
 from cacheroute.contracts.v1.common import utc_now
 from cacheroute.contracts.v1.errors import OutcomeCode
-from kdn_server.domain import CacheOperationState, CacheOperationTask, CacheOperationType
+from cacheroute.cache import CacheOperationState, CacheOperationTask, CacheOperationType
 from .base import GatewayAdapterBase
 from .capabilities import SupportState
