+++ src/cacheroute/routing/__init__.py
@@ -1,0 +1,5 @@
+"""Reusable routing domain state."""
+
+from .queue import QueueState, QueueWork
+
+__all__ = ["QueueState", "QueueWork"]
