+++ src/cacheroute/runtime/__init__.py
@@ -1,5 +1,6 @@
 """Dependency-light runtime identity and lifecycle vocabulary."""
 
 from .profiles import RuntimeProfile
+from .state import Snapshot, StateTransitionError, StrEnum
 
-__all__ = ["RuntimeProfile"]
+__all__ = ["RuntimeProfile", "Snapshot", "StateTransitionError", "StrEnum"]
