+++ src/cacheroute/topology/__init__.py
@@ -1,0 +1,5 @@
+"""Cluster topology and endpoint state."""
+
+from .lmcache import LMCacheEndpoint, LMCacheGatewayProfile
+
+__all__ = ["LMCacheEndpoint", "LMCacheGatewayProfile"]
